+++ src/api/Migrations/20240410125336_MigrateSections.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using BDMS;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using NetTopologySuite.Geometries;
 using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
@@ -12,9 +13,11 @@ using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
 namespace BDMS.Migrations
 {
     [DbContext(typeof(BdmsContext))]
-    partial class BdmsContextModelSnapshot : ModelSnapshot
+    [Migration("20240410125336_MigrateSections")]
+    partial class MigrateSections
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -2897,7 +2900,7 @@ namespace BDMS.Migrations
             modelBuilder.Entity("BDMS.Models.Section", b =>
                 {
                     b.HasOne("BDMS.Models.Borehole", "Borehole")
-                        .WithMany("Sections")
+                        .WithMany()
                         .HasForeignKey("BoreholeId")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
@@ -3095,8 +3098,6 @@ namespace BDMS.Migrations
                     b.Navigation("BoreholeFiles");
 
                     b.Navigation("Completions");
-
-                    b.Navigation("Sections");
 
                     b.Navigation("Stratigraphies");